+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Added
 
 ### Changed
+- 上映イベントの空席状況と、上映イベント自体のインポート期間を合わせるように調整。
 
 ### Deprecated